backend/requirements.txt
fastapi
pydantic
sqlmodel
uvicorn
transformers
torch
mysql-connector-python
python-clerk
